+++ module6-7/src/main/java/com/github/beguy/module6/accountType/AccountTypeInfoDto.java
@@ -2,9 +2,9 @@ package com.github.beguy.module6.accountType;
 
 public class AccountTypeInfoDto {
     private final AccountType accountType;
-    private final Integer clientsAmount;
+    private final Long clientsAmount;
 
-    AccountTypeInfoDto(AccountType accountType, Integer clientsAmount){
+    AccountTypeInfoDto(AccountType accountType, Long clientsAmount){
         this.accountType = accountType;
         this.clientsAmount = clientsAmount;
     }
@@ -13,7 +13,7 @@ public class AccountTypeInfoDto {
         return accountType;
     }
 
-    public Integer getClientsAmount() {
+    public Long getClientsAmount() {
         return clientsAmount;
     }
 }
